daemon/requirements.txt
PyYAML==5.1
gTTS==2.0.1
gtts-token==1.1.2
redis==2.10.6
mock==2.0.0
coverage==4.5.1
